pyproject.toml
[tool.poetry]
name = "fastapi-mongo"
version = "0.1.0"
description = ""
authors = ["Ilan Emanuel Fritzler <contacto.fritzlerilan@gmail.com>"]


packages = [
    { include="app", from="." },
]

[tool.poetry.scripts]
start = 'app.main:run_server'

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.78.0"}
uvicorn = "^0.17.6"
motor = "^3.0.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
